[FIX]  column에서 중복되는 부분을 하나로 묶고 그에 해당하는 파일 삭제
