--- poodle/src/components/Grade/VolanteerWorkTimeAttendance/Attendance/column/index.ts
@@ -1,4 +1,0 @@
-export { default as PerceptionColumn } from './PerceptionColumn';
-export { default as AbsentColumn } from './AbsentColumn';
-export { default as CutClassColumn } from './CutClassColumn';
-export { default as LeaveLateColumn } from './LeaveLateColumn';
